fix(F-417): add common.ts
