+++ frontend/src/interfaces/models/Common.ts
@@ -1,0 +1,6 @@
+export interface SessionScores {
+  clarity: number;
+  empathy: number;
+  structure: number;
+  focus: number;
+}
